// src/app/core/models/teacher-info.ts
export class TeacherInfo {

    id: number;
    firstName: string;
    lastName: string;
    fathersName: string;
    countryName: string;
    cityName: string;
    userId: number;
    userBirthday: Date;
    aboutMe: string;
    countryId: number;
    cityId: number;
    grade: number;
    gradeName: string;
    userQuantity: number;

}